feat: add cloudnative-pg and use it for harbor external database (#1120)

+++ upgrades/pre/upgrade-0-22-0.sh
@@ -2,11 +2,22 @@
 
 set -eu
 
-[[ ! $(helm status -n kube-system kured) ]] && echo "The old kured release does not exists. Skipping" && exit 0
-# It is safe to uninstall the release as it is stateless app.
-helm uninstall -n kube-system kured
+if [[ $(helm status -n harbor harbor) ]]; then
+  kubectl annotate -n harbor secret/harbor-database helm.sh/resource-policy='keep' deprecated=true
+  kubectl annotate -n harbor sts/harbor-database helm.sh/resource-policy='keep' deprecated=true
+  kubectl annotate -n harbor svc/harbor-database helm.sh/resource-policy='keep' deprecated=true
+else
+  echo "Harbor doesn't exist. Skipping"
+fi
 
+if [[ $(helm status -n kube-system kured) ]]; then
+  helm uninstall -n kube-system kured
+else
+  echo "The old kured release does not exists. Skipping"
+fi
 
-[[ ! $(helm status -n monitoring prometheus-operator) ]] && echo "The old prometheus-operator release does not exists. Skipping" && exit 0
-# It is safe to uninstall the release as it is stateless app.
-helm uninstall -n monitoring prometheus-operator
+if [[ $(helm status -n monitoring prometheus-operator) ]]; then
+  helm uninstall -n monitoring prometheus-operator
+else
+  echo "The old prometheus-operator release does not exists. Skipping"
+fi
